requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
aksharamukha==2.1.2
python-dotenv==1.0.1
httpx==0.25.2
pytest==7.4.4

